+++ tests/Integration/Application/Bus/MathCommandBus.php
@@ -10,12 +10,12 @@
 
 declare(strict_types=1);
 
-namespace CloudCreativity\Modules\Tests\Integration\Application;
+namespace CloudCreativity\Modules\Tests\Integration\Application\Bus;
 
 use CloudCreativity\Modules\Application\Bus\CommandDispatcher;
 use CloudCreativity\Modules\Application\Bus\Middleware\LogMessageDispatch;
-use CloudCreativity\Modules\Application\Bus\Through;
 use CloudCreativity\Modules\Application\Bus\WithCommand;
+use CloudCreativity\Modules\Application\Messages\Through;
 
 #[Through(LogMessageDispatch::class)]
 #[WithCommand(AddCommand::class, AddCommandHandler::class)]