+++ tests/Integration/Application/Bus/MultiplyCommandHandler.php
@@ -10,10 +10,10 @@
 
 declare(strict_types=1);
 
-namespace CloudCreativity\Modules\Tests\Integration\Application;
+namespace CloudCreativity\Modules\Tests\Integration\Application\Bus;
 
 use CloudCreativity\Modules\Application\Bus\Middleware\ExecuteInUnitOfWork;
-use CloudCreativity\Modules\Application\Bus\Through;
+use CloudCreativity\Modules\Application\Messages\Through;
 use CloudCreativity\Modules\Toolkit\Result\Result;
 
 #[Through(ExecuteInUnitOfWork::class)]